refactor: move agent runtime module root
