+++ tests/Unit/backend/web/services/test_chat_delivery_hook.py
@@ -14,11 +14,20 @@ from messaging.delivery.dispatcher import ChatDeliveryRequest
 def test_delivery_paths_depend_on_agent_runtime_port_not_native_gateway() -> None:
     delivery_source = inspect.getsource(chat_delivery_hook)
     threads_source = inspect.getsource(threads_router)
+    from backend.web.core import lifespan as lifespan_module
+
+    lifespan_source = inspect.getsource(lifespan_module)
 
     assert "NativeAgentRuntimeGateway" not in delivery_source
     assert "NativeAgentRuntimeGateway" not in threads_source
     assert "get_agent_runtime_gateway" in delivery_source
     assert "get_agent_runtime_gateway" in threads_source
+    assert "backend.agent_runtime.port" in delivery_source
+    assert "backend.agent_runtime.port" in threads_source
+    assert "backend.web.services.agent_runtime_port" not in delivery_source
+    assert "backend.web.services.agent_runtime_port" not in threads_source
+    assert "backend.agent_runtime.gateway" in lifespan_source
+    assert "backend.web.services.agent_runtime_gateway" not in lifespan_source
 
 
 @pytest.mark.asyncio
